+++ src/global/README.md
@@ -1,0 +1,1 @@
+The integration test uses items from this folder -- move out into common library
