+++ src/metrics/record.js
@@ -1,5 +1,3 @@
-const { prompt } = require('enquirer');
-
 const moment = require('moment');
 const debug = require('debug')('pgsh:metrics');
 const fs = require('fs');
@@ -10,6 +8,7 @@ const packageJson = JSON.parse(
   ),
 );
 
+const askForOptIn = require('./opt-in');
 const getCpuMetrics = require('./cpu');
 const config = require('../config');
 const global = require('../global');
@@ -18,21 +17,6 @@ const { METRICS_ENABLED } = require('../global/keys');
 const { get: getCommandLine } = require('./command-line');
 const { get: getStartedAt } = require('./timer');
 const store = require('./store');
-
-const askForOptIn = async () => {
-  const metricsEnabled = global.get(METRICS_ENABLED);
-  if (metricsEnabled === undefined) {
-    // TODO: ask the user to opt-in; write true or false
-    const { shouldEnable } = await prompt({
-      type: 'toggle',
-      name: 'shouldEnable',
-      message: 'Would you like to send anonymous usage data to support further pgsh development?',
-    });
-    global.set(METRICS_ENABLED, shouldEnable);
-    return shouldEnable;
-  }
-  return metricsEnabled;
-};
 
 const createSample = (exitCode) => ({
   ...getCpuMetrics(),
